lightroom-auto-system/select-mvp/backend/requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.12.0
pillow>=11.3.0
